+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2022-10-04
+### Fixed
+- `main.go` Exit code after showing a version.
+
+### Changed
+- `main.go` Catch both *SIGINT* and *SIGTERM* and exit with **0** code.
+
+## [3.0.0] - 2022-09-30
+### Changed
+- **-graphiteAddress** takes `hostname:port` pair.
+- Multiple senders allowed by repeating **-graphiteAddress** flag.
+
+### Removed
+- **-graphitePort** flag.
+
 ## [2.2.0] - 2022-06-07
 ### Changed
 - Init as go module.